pyproject.toml
[project]
name = "energy-forecast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlt[postgres]>=1.21.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
]

[project.scripts]
test = "pytest:main"
junior-dev = "energy_forecast.cli.junior_dev:main"

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]

[tool.ty.src]
include = ["src", "tests"]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "ty>=0.0.0",
]
